// resources/js/types/penjualan.ts
// Generated by https://quicktype.io

export interface Detail {
    id: number;
    penjualan_id: number;
    produk_id: number;
    qty: number;
    created_at: Date;
    updated_at: Date;
    produk: Produk;
    penjualan: Penjualan;
}

export interface Penjualan {
    id: number;
    invoice_number: InvoiceNumber;
    user_id: number;
    customer_id: number;
    type: Type;
    sub_total: number;
    ppn: number;
    biaya_layanan: number;
    total: number;
    laba: number;
    duitku_reference: null;
    status: Status;
    payment_url: null;
    created_at: Date;
    updated_at: Date;
    meja_id: number;
    payment_method: PaymentMethod;
    customer?: Customer;
    meja?: Meja;
    details?: Detail[];
    user?: Customer;
}

export interface Produk {
    id: number;
    kategori_id: number;
    satuan_id: number;
    nama_produk: string;
    slug: string;
    deskripsi: string;
    harga_beli: number;
    harga_jual: number;
    promo: number;
    percentage: number;
    stok: number;
    is_active: number;
    image: null | string;
    created_at: Date;
    updated_at: Date;
    kategori: Kategori;
    satuan: Satuan;
}

export interface Kategori {
    id: number;
    kategori: string;
    slug: string;
    icon: string;
    created_at: Date;
    updated_at: Date;
}

export interface Satuan {
    id: number;
    satuan: string;
    created_at: Date;
    updated_at: Date;
}

export interface Customer {
    id: number;
    name: string;
    email: string;
    email_verified_at: null;
    phone: null | string;
    address: null | string;
    avatar: null | string;
    role: string;
    is_active: number;
    created_at: Date;
    updated_at: Date;
}

export enum InvoiceNumber {
    Inv20250616Apkrpu = 'INV20250616APKRPU',
    Inv20250616P7Tjee = 'INV20250616P7TJEE',
    Inv20250616Quth7V = 'INV20250616QUTH7V',
}

export interface Meja {
    id: number;
    nama: string;
    status: string;
    kapasitas: number;
    created_at: Date;
    updated_at: Date;
}

export enum PaymentMethod {
    Cash = 'cash',
}

export enum Status {
    Paid = 'paid',
}

export enum Type {
    DineIn = 'dine_in',
}